move things around to clean up dependencies
