+++ services/api-server/tests/unit/test_api_files.py
@@ -1,19 +1,26 @@
+import datetime
+
 # pylint: disable=protected-access
 # pylint: disable=redefined-outer-name
 # pylint: disable=too-many-arguments
 # pylint: disable=unused-argument
 # pylint: disable=unused-variable
 from pathlib import Path
+from uuid import UUID
 
 import httpx
 import pytest
 from aioresponses import aioresponses as AioResponsesMock
 from fastapi import status
 from httpx import AsyncClient
-from models_library.api_schemas_storage import FileUploadCompleteLinks
+from models_library.api_schemas_storage import (
+    ETag,
+    FileUploadCompleteLinks,
+    FileUploadCompletionBody,
+    UploadedPart,
+)
 from pydantic import parse_obj_as
 from pytest_simcore.services_api_mocks_for_aiohttp_clients import (
-    DummyFileData,
     storage_v0_service_mock,
 )
 from respx import MockRouter
@@ -23,6 +30,41 @@ from simcore_service_api_server.models.schemas.files import ClientFileUploadSche
 pytest_plugins = [
     "pytest_simcore.aioresponses_mocker",
 ]
+
+
+class DummyFileData:
+    """Static class for providing consistent dummy file data for testing"""
+
+    _file_id: UUID = UUID("3fa85f64-5717-4562-b3fc-2c963f66afa6")
+    _file_name: str = "myfile.txt"
+    _final_e_tag: ETag = "07d1c1a4-b073-4be7-b022-f405d90e99aa"
+    _file_size: int = 100000
+
+    @classmethod
+    def file(cls) -> File:
+        return File(
+            id=File.create_id(
+                cls._file_size,
+                cls._file_name,
+                datetime.datetime.now(datetime.timezone.utc).isoformat(),
+            ),
+            filename=cls._file_name,
+            checksum="",
+        )
+
+    @classmethod
+    def file_size(cls) -> int:
+        return cls._file_size
+
+    @classmethod
+    def uploaded_parts(cls) -> FileUploadCompletionBody:
+        return FileUploadCompletionBody(
+            parts=[UploadedPart(number=ii + 1, e_tag=fake.uuid4()) for ii in range(5)]
+        )
+
+    @classmethod
+    def final_e_tag(cls) -> ETag:
+        return cls._final_e_tag
 
 
 @pytest.mark.xfail(reason="Under dev")
